stop trying to create an Elasticsearch cluster
If you are using ECK, you already have to install it yourself before
installing this chart, because it's a cluster-wide decision owing to
the CRDs it installs, so it's not something this chart should do
automatically. The flexibility required from this chart when using ECK
requires that the user be able to provide the entire Elasticsearch
custom resource's spec anyway. So there's no value added for the user
if this chart creates the Elasticsearch resource.

In addition, in the case where you want to share an Elasticsearch
between TheHive and Cortex, which chart should own it? Should it be
removed when one or the other chart is uninstalled? Nothing is simpler
if we create the Elasticsearch; everything is simpler if we don't.

If you are not using ECK, you already had to set up the Elasticsearch
server/cluster you are using anyway.

What remains is that if you are using ECK, this chart can make some
assumptions about the names of resources based on the name of the
preexisting Elasticsearch custom resource, simplifying configuration
of this chart.